Break up monolith into smaller files
