+++ case_conversion/__init__.py
@@ -1,8 +1,2 @@
-# from __future__ import absolute_import
-
-# from .case_conversion import (
-#     camelcase, pascalcase, snakecase, dashcase, kebabcase, spinalcase,
-#     constcase, screaming_snakecase, dotcase, separate_words, slashcase,
-#     backslashcase)
-
-from .case_conversion import Case, CaseConverter, InvalidAcronymError
+from .converter import CaseConverter
+from .types import Case, InvalidAcronymError
