--- case_conversion/case_conversion.py
@@ -1,558 +1,0 @@
-import sys
-from typing import cast, Any, Iterator, List, Optional, Tuple, Union
-import unicodedata
-from enum import Enum, unique
-
-from .alias import alias, aliased
-
-def _get_rubstring_ranges(a_str: str, sub: str) -> Iterator[Tuple[int, int]]:
-    start = 0
-    sub_len = len(sub)
-    while True:
-        start = a_str.find(sub, start)
-        if start == -1:
-            return
-        yield (start, start + sub_len)
-        start += 1
-
-
-def _char_is_sep(a_char: str) -> bool:
-    return not (
-        _char_is_upper(a_char) or _char_is_lower(a_char) or _char_is_decimal(a_char)
-    )
-
-
-def _char_is_decimal(a_char: str) -> bool:
-    return unicodedata.category(a_char) == "Nd"
-
-
-def _char_is_lower(a_char: str) -> bool:
-    return unicodedata.category(a_char) == "Ll"
-
-
-def _char_is_upper(a_char: str) -> bool:
-    return unicodedata.category(a_char) == "Lu"
-
-
-def _is_upper(a_string: str) -> bool:
-    return len(a_string) == 1 and _char_is_upper(a_string)
-
-
-def _is_valid_acronym(a_string: str) -> bool:
-    if len(a_string) == 0:
-        return False
-
-    for a_char in a_string:
-        if _char_is_sep(a_char):
-            return False
-
-    return True
-
-
-class InvalidAcronymError(Exception):
-    """Raise when acronym fails validation."""
-
-    def __init__(self, acronym: str) -> None:
-        msg = "Case Conversion: acronym '{}' is invalid.".format(acronym)
-        super().__init__(msg)
-
-
-@unique
-class Case(Enum):
-    # - upper: All words are upper-case.
-    # - lower: All words are lower-case.
-    # - pascal: All words are title-case or upper-case. Note that the
-    #           stringiable may still have separators.
-    # - camel: First word is lower-case, the rest are title-case or
-    #          upper-case. stringiable may still have separators.
-    # - mixed: Any other mixing of word casing. Never occurs if there are
-    #          no separators.
-    # - unknown: stringiable contains no words.
-    UNKOWN = "unknown"
-    UPPER = "upper"
-    LOWER = "lower"
-    CAMEL = "camel"
-    PASCAL = "pascal"
-    MIXED = "mixed"
-
-
-@aliased
-class CaseConverter(object):
-    """Main Class."""
-
-    @staticmethod
-    def _determine_case(was_upper: bool, words: List[str], string: str) -> Case:
-        """
-        Determine case type of string.
-
-        Arguments:
-            was_upper {[type]} -- [description]
-            words {[type]} -- [description]
-            string {[type]} -- [description]
-
-        Returns:
-            - upper: All words are upper-case.
-            - lower: All words are lower-case.
-            - pascal: All words are title-case or upper-case. Note that the
-                      stringiable may still have separators.
-            - camel: First word is lower-case, the rest are title-case or
-                     upper-case. stringiable may still have separators.
-            - mixed: Any other mixing of word casing. Never occurs if there are
-                     no separators.
-            - unknown: stringiable contains no words.
-
-        """
-        case_type = Case.UNKOWN
-        if was_upper:
-            case_type = Case.UPPER
-        elif string.islower():
-            case_type = Case.LOWER
-        elif words:
-            camel_case = words[0].islower()
-            pascal_case = words[0].istitle() or words[0].isupper()
-
-            if camel_case or pascal_case:
-                for word in words[1:]:
-                    c = word.istitle() or word.isupper()
-                    camel_case &= c
-                    pascal_case &= c
-                    if not c:
-                        break
-
-            if camel_case:
-                case_type = Case.CAMEL
-            elif pascal_case:
-                case_type = Case.PASCAL
-            else:
-                case_type = Case.MIXED
-
-        return case_type
-
-    @staticmethod
-    def _advanced_acronym_detection(
-        s: int, i: int, words: List[str], acronyms: List[str]
-    ) -> int:
-        """Detect acronyms by checking against a list of acronyms.
-
-        Check a run of words represented by the range [s, i].
-        Return last index of new word groups.
-        """
-        # Combine each letter into single string.
-        acstr = "".join(words[s:i])
-
-        # List of ranges representing found acronyms.
-        range_list: List[Tuple[int, int]] = []
-        # Set of remaining letters.
-        not_range = set(range(len(acstr)))
-
-        # Search for each acronym in acstr.
-        for acronym in acronyms:
-            for (a, b) in _get_rubstring_ranges(acstr, acronym):
-                # Make sure found acronym doesn't overlap with others.
-                ok = True
-                for r in range_list:
-                    if a < r[1] and b > r[0]:
-                        ok = False
-                        break
-
-                if ok:
-                    range_list.append((a, b))
-                    for j in range(a, b):
-                        not_range.remove(j)
-
-        # Add remaining letters as ranges.
-        for nr in not_range:
-            range_list.append((nr, nr + 1))
-
-        # No ranges will overlap, so it's safe to sort by lower bound,
-        # which sort() will do by default.
-        range_list.sort()
-
-        # Remove original letters in word list.
-        for _ in range(s, i):
-            del words[s]
-
-        # Replace them with new word grouping.
-        for j in range(len(range_list)):
-            r = range_list[j]
-            words.insert(s + j, acstr[r[0] : r[1]])
-
-        return s + len(range_list) - 1
-
-    @staticmethod
-    def _simple_acronym_detection(s: int, i: int, words: List[str], *args) -> int:
-        """Detect acronyms based on runs of upper-case letters."""
-        # Combine each letter into a single string.
-        acstr = "".join(words[s:i])
-
-        # Remove original letters in word list.
-        for _ in range(s, i):
-            del words[s]
-
-        # Replace them with new word grouping.
-        words.insert(s, "".join(acstr))
-
-        return s
-
-    @staticmethod
-    def _sanitize_acronyms(
-        unsafe_acronyms: Union[List[str], Tuple[str], Tuple[str, str], str]
-    ) -> List[str]:
-        """Normalize valid acronyms to upper-case.
-
-        If an invalid acronym is encountered (contains separators)
-        raise InvalidAcronymError.
-        """
-        acronyms = []
-        for a in unsafe_acronyms:
-            if _is_valid_acronym(a):
-                acronyms.append(a.upper())
-            else:
-                raise InvalidAcronymError(a)
-        return acronyms
-
-    @staticmethod
-    def _normalize_words(
-        words: List[str], acronyms: Union[List[str], str]
-    ) -> List[str]:
-        """Normalize case of each word to PascalCase."""
-        for i, _ in enumerate(words):  # TODO: Rewrite loop without enumerate
-            # if detect_acronyms:
-            if words[i].upper() in acronyms:
-                # Convert known acronyms to upper-case.
-                words[i] = words[i].upper()
-            else:
-                # Fallback behavior: Preserve case on upper-case words.
-                if not words[i].isupper():
-                    words[i] = words[i].capitalize()
-        return words
-
-    @classmethod
-    def _segment_string(cls, string: str) -> Tuple[List[Optional[str]], str, bool]:
-        """Segment string on separator into list of words.
-
-        Arguments:
-            string -- the string we want to process
-        Returns:
-            words -- list of words the string got minced to
-            separator -- the separator char intersecting words
-            was_upper -- whether string happened to be upper-case
-        """
-        words: List[Optional[str]] = []
-        separator = ""
-
-        # curr_index of current character. Initially 1 because we don't
-        # want to check if the 0th character is a boundary.
-        curr_i = 1
-        # Index of first character in a sequence
-        seq_i = 0
-        # Previous character.
-        prev_i = string[0:1]
-
-        # Treat an all-caps stringiable as lower-case, to prevent its
-        # letters to be counted as boundaries
-        was_upper = False
-        if string.isupper():
-            string = string.lower()
-            was_upper = True
-
-        # Iterate over each character, checking for boundaries, or places
-        # where the stringiable should divided.
-        while curr_i <= len(string):
-            char = string[curr_i : curr_i + 1]
-            split = False
-            if curr_i < len(string):
-                # Detect upper-case letter as boundary.
-                if _char_is_upper(char):
-                    split = True
-                # Detect transition from separator to not separator.
-                elif not _char_is_sep(char) and _char_is_sep(prev_i):
-                    split = True
-                # Detect transition not separator to separator.
-                elif _char_is_sep(char) and not _char_is_sep(prev_i):
-                    split = True
-            else:
-                # The looprev_igoes one extra iteration so that it can
-                # handle the remaining text after the last boundary.
-                split = True
-
-            if split:
-                if not _char_is_sep(prev_i):
-                    words.append(string[seq_i:curr_i])
-                else:
-                    # stringiable contains at least one separator.
-                    # Use the first one as the stringiable's primary separator.
-                    if not separator:
-                        separator = string[seq_i : seq_i + 1]
-
-                    # Use None to indicate a separator in the word list.
-                    words.append(None)
-                    # If separators weren't included in the list, then breaks
-                    # between upper-case sequences ("AAA_BBB") would be
-                    # disregarded; the letter-run detector would count them
-                    # as a single sequence ("AAABBB").
-                seq_i = curr_i
-
-            curr_i += 1
-            prev_i = char
-
-        return words, separator, was_upper
-
-    @classmethod
-    def parse_case(
-        cls,
-        string: str,
-        acronyms: Optional[List[str]] = None,
-        preserve_case: bool = False,
-    ) -> Tuple[List[str], Case, str]:
-        """
-        Parse a stringiable into a list of words.
-
-        Also returns the case type, which can be one of the following:
-            - upper: All words are upper-case.
-            - lower: All words are lower-case.
-            - pascal: All words are title-case or upper-case. Note that the
-                      stringiable may still have separators.
-            - camel: First word is lower-case, the rest are title-case or
-                     upper-case. stringiable may still have separators.
-            - mixed: Any other mixing of word casing. Never occurs if there are
-                     no separators.
-            - unknown: stringiable contains no words.
-
-        Also returns the first separator character,
-        or False if there isn't one.
-        """
-        words_with_sep, separator, was_upper = cls._segment_string(string)
-
-        if acronyms:
-            # Use advanced acronym detection with list
-            acronyms = cls._sanitize_acronyms(acronyms)
-            check_acronym = cls._advanced_acronym_detection  # type: ignore
-        else:
-            acronyms = []
-            # Fallback to simple acronym detection.
-            check_acronym = cls._simple_acronym_detection  # type: ignore
-
-        # Letter-run detector
-
-        # Index of current word.
-        i = 0
-        # Index of first letter in run.
-        s = None
-
-        # Find runs of single upper-case letters.
-        while i < len(words_with_sep):
-            word = words_with_sep[i]
-            if word is not None and _is_upper(word):
-                if s is None:
-                    s = i
-            elif s is not None:
-                i = check_acronym(s, i, words_with_sep, acronyms) + 1  # type: ignore
-                s = None
-            i += 1
-
-        # Separators are no longer needed, so they should be removed.
-        words: List[str] = [w for w in words_with_sep if w is not None]
-
-        # Determine case type.
-        case_type = cls._determine_case(was_upper, words, string)
-
-        if preserve_case:
-            if was_upper:
-                words = [w.upper() for w in words]
-        else:
-            words = cls._normalize_words(words, acronyms)
-
-        return words, case_type, separator
-
-    @classmethod
-    def camel(cls, text: str, acronyms: Optional[List[str]] = None) -> str:
-        """Return text in camelCase style.
-
-        Args:
-            text: input string to convert case
-            detect_acronyms: should attempt to detect acronyms
-            acronyms: a list of acronyms to detect
-
-        >>> camelcase("hello world")
-        'helloWorld'
-        >>> camelcase("HELLO_HTML_WORLD", True, ["HTML"])
-        'helloHTMLWorld'
-        """
-        words, _case, _sep = cls.parse_case(text, acronyms)
-        if words:
-            words[0] = words[0].lower()
-        return "".join(words)
-
-    @alias("mixed")
-    @classmethod
-    def pascal(cls, text: str, acronyms: Optional[List[str]] = None) -> str:
-        """Return text in PascalCase style (aka MixedCase).
-
-        Args:
-            text: input string to convert case
-            detect_acronyms: should attempt to detect acronyms
-            acronyms: a list of acronyms to detect
-
-        >>> pascalcase("hello world")
-        'HelloWorld'
-        >>> pascalcase("HELLO_HTML_WORLD", True, ["HTML"])
-        'HelloHTMLWorld'
-        """
-        words, _case, _sep = cls.parse_case(text, acronyms)
-        return "".join(words)
-
-    @classmethod
-    def snake(cls, text: str, acronyms: Optional[List[str]] = None) -> str:
-        """Return text in snake_case style.
-
-        Args:
-            text: input string to convert case
-            detect_acronyms: should attempt to detect acronyms
-            acronyms: a list of acronyms to detect
-
-        >>> snakecase("hello world")
-        'hello_world'
-        >>> snakecase("HelloHTMLWorld", True, ["HTML"])
-        'hello_html_world'
-        """
-        words, _case, _sep = cls.parse_case(text, acronyms)
-        return "_".join([w.lower() for w in words])
-
-    @alias("kebap", "spinal", "slug")
-    @classmethod
-    def dash(cls, text: str, acronyms: Optional[List[str]] = None) -> str:
-        """Return text in dash-case style (aka kebab-case, spinal-case).
-
-        Args:
-            text: input string to convert case
-            detect_acronyms: should attempt to detect acronyms
-            acronyms: a list of acronyms to detect
-
-        >>> dashcase("hello world")
-        'hello-world'
-        >>> dashcase("HelloHTMLWorld", True, ["HTML"])
-        'hello-html-world'
-        """
-        words, _case, _sep = cls.parse_case(text, acronyms)
-        return "-".join([w.lower() for w in words])
-
-    @alias("screaming")
-    @classmethod
-    def const(cls, text: str, acronyms: Optional[List[str]] = None) -> str:
-        """Return text in CONST_CASE style (aka SCREAMING_SNAKE_CASE).
-
-        Args:
-            text: input string to convert case
-            detect_acronyms: should attempt to detect acronyms
-            acronyms: a list of acronyms to detect
-
-        >>> constcase("hello world")
-        'HELLO_WORLD'
-        >>> constcase("helloHTMLWorld", True, ["HTML"])
-        'HELLO_HTML_WORLD'
-        """
-        words, _case, _sep = cls.parse_case(text, acronyms)
-        return "_".join([w.upper() for w in words])
-
-    @classmethod
-    def dot(cls, text: str, acronyms: Optional[List[str]] = None) -> str:
-        """Return text in dot.case style.
-
-        Args:
-            text: input string to convert case
-            detect_acronyms: should attempt to detect acronyms
-            acronyms: a list of acronyms to detect
-
-        >>> dotcase("hello world")
-        'hello.world'
-        >>> dotcase("helloHTMLWorld", True, ["HTML"])
-        'hello.html.world'
-        """
-        words, _case, _sep = cls.parse_case(text, acronyms)
-        return ".".join([w.lower() for w in words])
-
-    @classmethod
-    def separate_words(cls, text: str, acronyms: Optional[List[str]] = None) -> str:
-        """Return text in "seperate words" style.
-
-        Args:
-            text: input string to convert case
-            detect_acronyms: should attempt to detect acronyms
-            acronyms: a list of acronyms to detect
-
-        >>> separate_words("HELLO_WORLD")
-        'HELLO WORLD'
-        >>> separate_words("helloHTMLWorld", True, ["HTML"])
-        'hello HTML World'
-        """
-        words, _case, _sep = cls.parse_case(text, acronyms, preserve_case=True)
-        return " ".join(words)
-
-    @classmethod
-    def slash(cls, text: str, acronyms: Optional[List[str]] = None) -> str:
-        """Return text in slash/case style.
-
-        Args:
-            text: input string to convert case
-            detect_acronyms: should attempt to detect acronyms
-            acronyms: a list of acronyms to detect
-
-        >>> slashcase("HELLO_WORLD")
-        'HELLO/WORLD'
-        >>> slashcase("helloHTMLWorld", True, ["HTML"])
-        'hello/HTML/World'
-        """
-        words, _case, _sep = cls.parse_case(text, acronyms, preserve_case=True)
-        return "/".join(words)
-
-    @classmethod
-    def backslash(cls, text: str, acronyms: Optional[List[str]] = None) -> str:
-        r"""Return text in backslash\case style.
-
-        Args:
-            text: input string to convert case
-            detect_acronyms: should attempt to detect acronyms
-            acronyms: a list of acronyms to detect
-
-        >>> backslashcase("HELLO_WORLD") == r'HELLO\WORLD'
-        True
-        >>> backslashcase("helloHTMLWorld",
-            True, ["HTML"]) == r'hello\HTML\World'
-        True
-        """
-        words, _case, _sep = cls.parse_case(text, acronyms, preserve_case=True)
-        return "\\".join(words)
-
-    @alias("camel_snake")
-    @classmethod
-    def ada(cls, text: str, acronyms: Optional[List[str]] = None) -> str:
-        r"""Return text in Ada_Case style."""
-        words, _case, _sep = cls.parse_case(text, acronyms)
-        return "_".join([w.capitalize() for w in words])
-
-    @classmethod
-    def title(cls, text: str, acronyms: Optional[List[str]] = None) -> str:
-        r"""Return text in Title_case style."""
-        return cls.snake(text, acronyms).capitalize()
-
-    @classmethod
-    def lower(cls, text: str, acronyms: Optional[List[str]] = None) -> str:
-        r"""Return text in lowercase style."""
-        return text.lower()
-
-    @classmethod
-    def upper(cls, text: str, acronyms: Optional[List[str]] = None) -> str:
-        r"""Return text in UPPERCASE style."""
-        return text.upper()
-
-    @classmethod
-    def capital(cls, text: str, acronyms: Optional[List[str]] = None) -> str:
-        r"""Return text in UPPERCASE style."""
-        return text.capitalize()
-
-    @classmethod
-    def http_header(cls, text: str, acronyms: Optional[List[str]] = None) -> str:
-        r"""Return text in Http-Header-Case style."""
-        words, _case, _sep = cls.parse_case(text, acronyms)
-        return "-".join([w.capitalize() for w in words])
