+++ tests/test_internals.py
@@ -1,6 +1,8 @@
 import pytest
 
 from case_conversion import Case, CaseConverter, InvalidAcronymError
+import case_conversion.utils as utils
+
 
 @pytest.mark.parametrize(
     "string,expected",
@@ -16,10 +18,10 @@ from case_conversion import Case, CaseConverter, InvalidAcronymError
         ("foo\\bar\\string", (["foo", None, "bar", None, "string"], "\\", False)),
         ("foobarstring", (["foobarstring"], "", False)),
         ("FOOBARSTRING", (["foobarstring"], "", True)),
-    )
+    ),
 )
 def test_segment_string(string, expected):
-    assert CaseConverter._segment_string(string) == expected
+    assert utils._segment_string(string) == expected
 
 
 @pytest.mark.parametrize(
@@ -30,10 +32,10 @@ def test_segment_string(string, expected):
         (("Http",), ["HTTP"],),
         (("httP",), ["HTTP"],),
         (("http", "Nasa"), ["HTTP", "NASA"]),
-    )
+    ),
 )
 def test_sanitize_acronyms(acronyms, expected):
-    assert CaseConverter._sanitize_acronyms(acronyms) == expected
+    assert utils._sanitize_acronyms(acronyms) == expected
 
 
 @pytest.mark.parametrize(
@@ -42,10 +44,10 @@ def test_sanitize_acronyms(acronyms, expected):
         # TODO: Add more cases
         (0, 1, ["FOO", "bar"], 0),
         (1, 2, ["foo", "BAR", "baz"], 1),
-    )
+    ),
 )
 def test_simple_acronym_detection(s, i, words, expected):
-    assert CaseConverter._simple_acronym_detection(s, i, words) == expected
+    assert utils._simple_acronym_detection(s, i, words) == expected
 
 
 @pytest.mark.parametrize(
@@ -54,10 +56,10 @@ def test_simple_acronym_detection(s, i, words, expected):
         # TODO: Add more cases
         (0, 1, ["FOO", "bar"], ("FOO",), 0),
         (0, 1, ["FOO", "bar"], ("BAR",), 2),
-    )
+    ),
 )
 def test_advanced_acronym_detection(s, i, words, acronyms, expected):
-    assert CaseConverter._advanced_acronym_detection(s, i, words, acronyms) == expected
+    assert utils._advanced_acronym_detection(s, i, words, acronyms) == expected
 
 
 @pytest.mark.parametrize(
@@ -67,7 +69,7 @@ def test_advanced_acronym_detection(s, i, words, acronyms, expected):
         ("fooBarBaz", None, True, (["foo", "Bar", "Baz"], Case.CAMEL, "")),
         ("fooBarBaz", ("BAR",), False, (["Foo", "BAR", "Baz"], Case.CAMEL, "")),
         ("fooBarBaz", ("BAR",), True, (["foo", "Bar", "Baz"], Case.CAMEL, "")),
-    )
+    ),
 )
 def test_parse_case(string, acronyms, preserve_case, expected):
     assert CaseConverter.parse_case(string, acronyms, preserve_case) == expected
@@ -82,15 +84,10 @@ def test_invalid_acronym_error_message():
         assert msg in str(e)
 
 
-@pytest.mark.parametrize(
-    "acronyms",
-    (
-        "HT-TP", "NA SA", "SU.GAR"
-    )
-)
+@pytest.mark.parametrize("acronyms", ("HT-TP", "NA SA", "SU.GAR"))
 def test_sanitize_acronyms_raises_on_invalid_acronyms(acronyms):
     with pytest.raises(InvalidAcronymError):
-        CaseConverter._sanitize_acronyms(acronyms)
+        utils._sanitize_acronyms(acronyms)
 
 
 @pytest.mark.parametrize(
@@ -100,10 +97,10 @@ def test_sanitize_acronyms_raises_on_invalid_acronyms(acronyms):
         (["fooBar"], (), ["Foobar"]),
         (["FooBar"], (), ["Foobar"]),
         (["Foo", "Bar"], ("BAR"), ["Foo", "BAR"]),
-    )
+    ),
 )
 def test_normalize_words(words, acronyms, expected):
-    assert CaseConverter._normalize_words(words, acronyms) == expected
+    assert utils._normalize_words(words, acronyms) == expected
 
 
 @pytest.mark.parametrize(
@@ -115,7 +112,7 @@ def test_normalize_words(words, acronyms, expected):
         (False, ["foo", "Bar"], "", Case.CAMEL),
         (False, ["Foo", "Bar"], "", Case.PASCAL),
         (False, ["foo", "bar"], "", Case.MIXED),
-    )
+    ),
 )
 def test_determine_case(was_upper, words, string, expected):
-    assert CaseConverter._determine_case(was_upper, words, string) == expected
+    assert utils._determine_case(was_upper, words, string) == expected
